restructuring unit tests
