+++ tests/HydraScript.UnitTests/Domain/BackEnd/CallTests.cs
@@ -1,8 +1,7 @@
 using HydraScript.Domain.BackEnd;
 using HydraScript.Domain.BackEnd.Impl.Addresses;
-using Xunit;
 
-namespace HydraScript.UnitTests.Unit.BackEnd;
+namespace HydraScript.UnitTests.Domain.BackEnd;
 
 public class CallTests
 {
