+++ tests/HydraScript.UnitTests/Domain/BackEnd/HashAddressTests.cs
@@ -1,7 +1,6 @@
 using HydraScript.Domain.BackEnd.Impl.Addresses;
-using Xunit;
 
-namespace HydraScript.UnitTests.Unit.BackEnd;
+namespace HydraScript.UnitTests.Domain.BackEnd;
 
 public class HashAddressTests
 {
