+++ tests/HydraScript.UnitTests/Domain/BackEnd/ValuesTests.cs
@@ -1,7 +1,6 @@
 using HydraScript.Domain.BackEnd.Impl.Values;
-using Xunit;
 
-namespace HydraScript.UnitTests.Unit.BackEnd;
+namespace HydraScript.UnitTests.Domain.BackEnd;
 
 public class ValuesTests
 {
