+++ tests/HydraScript.UnitTests/Domain/FrontEnd/ParserData.cs
@@ -1,6 +1,6 @@
 using System.Collections;
 
-namespace HydraScript.UnitTests.TestData;
+namespace HydraScript.UnitTests.Domain.FrontEnd;
 
 public class ParserSuccessTestData : IEnumerable<object[]>
 {
