+++ tests/HydraScript.UnitTests/Domain/FrontEnd/RegexLexerTests.cs
@@ -2,10 +2,8 @@ using HydraScript.Domain.FrontEnd.Lexer;
 using HydraScript.Domain.FrontEnd.Lexer.Impl;
 using HydraScript.Domain.FrontEnd.Lexer.TokenTypes;
 using HydraScript.Infrastructure;
-using HydraScript.UnitTests.TestData;
-using Xunit;
 
-namespace HydraScript.UnitTests.Unit.FrontEnd;
+namespace HydraScript.UnitTests.Domain.FrontEnd;
 
 public class RegexLexerTests
 {
