+++ tests/HydraScript.UnitTests/Domain/FrontEnd/StructureTests.cs
@@ -2,9 +2,8 @@ using HydraScript.Domain.FrontEnd.Lexer;
 using HydraScript.Domain.FrontEnd.Lexer.Impl;
 using HydraScript.Domain.FrontEnd.Lexer.TokenTypes;
 using HydraScript.Infrastructure;
-using Xunit;
 
-namespace HydraScript.UnitTests.Unit.FrontEnd;
+namespace HydraScript.UnitTests.Domain.FrontEnd;
 
 public class StructureTests
 {
