+++ tests/HydraScript.UnitTests/Domain/IR/ObjectTypeTests.cs
@@ -1,7 +1,6 @@
 using HydraScript.Domain.IR.Types;
-using Xunit;
 
-namespace HydraScript.UnitTests.Unit.IR.Types;
+namespace HydraScript.UnitTests.Domain.IR;
 
 public class ObjectTypeTests
 {
