+++ tests/HydraScript.UnitTests/GlobalUsings.cs
@@ -1,5 +1,6 @@
 // Global using directives
 
+global using Xunit;
 global using FluentAssertions;
 global using Type =  HydraScript.Domain.IR.Types.Type;
 global using NSubstitute;
