+++ tests/HydraScript.UnitTests/Properties/AssemblyInfo.cs
@@ -1,2 +1,1 @@
-using Xunit;
 [assembly: AssemblyTrait("Category", "Unit")]
